requirements.txt
Werkzeug==1.0.1
Flask==1.1.2
filehash==0.1.dev5
requests==2.25.1
ioc_finder==5.0.1
iocextract==1.13.1
ipwhois==1.2.0
pandas==1.2.3
pefile==2019.4.18
plotly==4.14.3
python_magic==0.4.22
python-magic-bin==0.4.14
scapy==2.4.4
valhallaAPI==0.5.1
